#59 Move Cleaner::cleanUp() method to the DrupalProject class
